<refactor>: RandomNumber 클래스 명을 NumberGenerator로 수정
