+++ src/main/java/lotto/util/NumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.util;
 
 import java.util.List;
